requirements.txt
Appium-Python-Client>=4.0.0
pytest>=8.2.2
pytest-bdd>=7.2.0
pytest-html>=4.1.1